Move GroupKeys to c.s.client.stream package
