+++ client/src/main/java/com/streamr/client/Streamr.java
@@ -12,7 +12,7 @@ import com.streamr.client.rest.StreamrRestClient;
 import com.streamr.client.rest.UserInfo;
 import com.streamr.client.subs.Subscription;
 import com.streamr.client.utils.Address;
-import com.streamr.client.utils.GroupKey;
+import com.streamr.client.stream.GroupKey;
 import java.io.IOException;
 import java.math.BigInteger;
 import java.util.Date;
