+++ client/src/main/java/com/streamr/client/StreamrClient.java
@@ -25,7 +25,6 @@ import com.streamr.client.protocol.message_layer.AbstractGroupKeyMessage;
 import com.streamr.client.protocol.message_layer.GroupKeyRequest;
 import com.streamr.client.protocol.message_layer.MalformedMessageException;
 import com.streamr.client.protocol.message_layer.StreamMessage;
-import com.streamr.client.protocol.message_layer.StreamMessageValidator;
 import com.streamr.client.rest.AmbiguousResultsException;
 import com.streamr.client.rest.DataUnionSecretResponse;
 import com.streamr.client.rest.Permission;
@@ -41,11 +40,11 @@ import com.streamr.client.subs.RealTimeSubscription;
 import com.streamr.client.subs.Subscription;
 import com.streamr.client.utils.Address;
 import com.streamr.client.utils.AddressValidityUtil;
-import com.streamr.client.utils.GroupKey;
-import com.streamr.client.utils.GroupKeyStore;
+import com.streamr.client.stream.GroupKey;
+import com.streamr.client.stream.GroupKeyStore;
 import com.streamr.client.utils.IdGenerator;
-import com.streamr.client.utils.InMemoryGroupKeyStore;
-import com.streamr.client.utils.KeyExchangeUtil;
+import com.streamr.client.stream.InMemoryGroupKeyStore;
+import com.streamr.client.stream.KeyExchangeUtil;
 import com.streamr.client.utils.MessageCreationUtil;
 import com.streamr.client.utils.OneTimeResend;
 import java.io.IOException;
