+++ client/src/main/java/com/streamr/client/protocol/message_layer/EncryptedGroupKeyAdapter.java
@@ -3,7 +3,7 @@ package com.streamr.client.protocol.message_layer;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
-import com.streamr.client.utils.EncryptedGroupKey;
+import com.streamr.client.stream.EncryptedGroupKey;
 
 import javax.annotation.Nullable;
 import java.io.IOException;
