+++ client/src/main/java/com/streamr/client/protocol/message_layer/GroupKeyAnnounceAdapter.java
@@ -2,7 +2,7 @@ package com.streamr.client.protocol.message_layer;
 
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
-import com.streamr.client.utils.EncryptedGroupKey;
+import com.streamr.client.stream.EncryptedGroupKey;
 import java.io.IOException;
 import java.util.List;
 import javax.annotation.Nullable;
