+++ client/src/main/java/com/streamr/client/stream/GroupKeyStore.java
@@ -1,4 +1,4 @@
-package com.streamr.client.utils;
+package com.streamr.client.stream;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
