+++ client/src/main/java/com/streamr/client/subs/BasicSubscription.java
@@ -1,16 +1,16 @@
 package com.streamr.client.subs;
 
 import com.streamr.client.MessageHandler;
+import com.streamr.client.stream.EncryptionUtil;
 import com.streamr.client.utils.UnableToDecryptException;
 import com.streamr.client.protocol.common.MessageRef;
 import com.streamr.client.protocol.common.UnsupportedMessageException;
 import com.streamr.client.protocol.message_layer.StreamMessage;
 import com.streamr.client.utils.Address;
 import com.streamr.client.utils.DecryptionQueues;
-import com.streamr.client.utils.EncryptionUtil;
-import com.streamr.client.utils.GroupKey;
-import com.streamr.client.utils.GroupKeyStore;
-import com.streamr.client.utils.KeyExchangeUtil;
+import com.streamr.client.stream.GroupKey;
+import com.streamr.client.stream.GroupKeyStore;
+import com.streamr.client.stream.KeyExchangeUtil;
 import com.streamr.client.utils.OrderedMsgChain;
 import com.streamr.client.utils.OrderingUtil;
 import java.util.ArrayList;
