+++ client/src/main/java/com/streamr/client/subs/HistoricalSubscription.java
@@ -5,9 +5,9 @@ import com.streamr.client.options.ResendOption;
 import com.streamr.client.protocol.common.UnsupportedMessageException;
 import com.streamr.client.protocol.message_layer.StreamMessage;
 import com.streamr.client.utils.Address;
-import com.streamr.client.utils.GroupKey;
-import com.streamr.client.utils.GroupKeyStore;
-import com.streamr.client.utils.KeyExchangeUtil;
+import com.streamr.client.stream.GroupKey;
+import com.streamr.client.stream.GroupKeyStore;
+import com.streamr.client.stream.KeyExchangeUtil;
 import java.util.Collection;
 import java.util.function.Consumer;
 import org.slf4j.Logger;
