+++ client/src/main/java/com/streamr/client/subs/RealTimeSubscription.java
@@ -4,6 +4,8 @@ import com.streamr.client.MessageHandler;
 import com.streamr.client.options.ResendOption;
 import com.streamr.client.protocol.common.UnsupportedMessageException;
 import com.streamr.client.protocol.message_layer.StreamMessage;
+import com.streamr.client.stream.GroupKeyStore;
+import com.streamr.client.stream.KeyExchangeUtil;
 import com.streamr.client.utils.*;
 import java.util.ArrayList;
 import org.slf4j.Logger;
