+++ client/src/main/java/com/streamr/client/utils/MessageCreationUtil.java
@@ -14,6 +14,13 @@ import com.streamr.client.protocol.message_layer.MessageId;
 import com.streamr.client.protocol.message_layer.StreamMessage;
 import com.streamr.client.protocol.message_layer.StringOrMillisDateJsonAdapter;
 import com.streamr.client.rest.Stream;
+import com.streamr.client.stream.EncryptedGroupKey;
+import com.streamr.client.stream.EncryptionUtil;
+import com.streamr.client.stream.GroupKey;
+import com.streamr.client.stream.InvalidGroupKeyException;
+import com.streamr.client.stream.InvalidGroupKeyRequestException;
+import com.streamr.client.stream.InvalidGroupKeyResponseException;
+import com.streamr.client.stream.KeyExchangeUtil;
 import java.math.BigInteger;
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
