+++ client/src/test/groovy/com/streamr/client/SubscriptionsSpec.groovy
@@ -2,8 +2,8 @@ package com.streamr.client
 
 import com.streamr.client.subs.RealTimeSubscription
 import com.streamr.client.subs.Subscription
-import com.streamr.client.utils.GroupKeyStore
-import com.streamr.client.utils.KeyExchangeUtil
+import com.streamr.client.stream.GroupKeyStore
+import com.streamr.client.stream.KeyExchangeUtil
 import java.util.function.Consumer
 import spock.lang.Specification
