+++ client/src/test/groovy/com/streamr/client/subs/CombinedSubscriptionSpec.groovy
@@ -9,8 +9,8 @@ import com.streamr.client.testing.TestingAddresses
 import com.streamr.client.testing.TestingContent
 import com.streamr.client.testing.TestingMessageRef
 import com.streamr.client.utils.Address
-import com.streamr.client.utils.GroupKeyStore
-import com.streamr.client.utils.KeyExchangeUtil
+import com.streamr.client.stream.GroupKeyStore
+import com.streamr.client.stream.KeyExchangeUtil
 import com.streamr.client.utils.OrderedMsgChain
 import spock.lang.Specification
