+++ client/src/test/groovy/com/streamr/client/utils/DecryptionQueuesSpec.groovy
@@ -2,6 +2,8 @@ package com.streamr.client.utils
 
 import com.streamr.client.protocol.message_layer.MessageId
 import com.streamr.client.protocol.message_layer.StreamMessage
+import com.streamr.client.stream.EncryptionUtil
+import com.streamr.client.stream.GroupKey
 import com.streamr.client.testing.TestingAddresses
 import com.streamr.client.testing.TestingContent
 import com.streamr.client.testing.TestingMessageRef
