+++ client/src/test/groovy/com/streamr/client/utils/MessageCreationUtilSpec.groovy
@@ -9,6 +9,10 @@ import com.streamr.client.protocol.message_layer.GroupKeyRequest
 import com.streamr.client.protocol.message_layer.GroupKeyResponse
 import com.streamr.client.protocol.message_layer.StreamMessage
 import com.streamr.client.rest.Stream
+import com.streamr.client.stream.EncryptionUtil
+import com.streamr.client.stream.GroupKey
+import com.streamr.client.stream.InvalidGroupKeyRequestException
+import com.streamr.client.stream.KeyExchangeUtil
 import com.streamr.client.testing.TestingAddresses
 import spock.lang.Specification
